ARM: omap4: tuna: Fix warm reset
After a warm reset, gpios switch their I/O state to match the
reset state, but not their pullup state.  This caused the
eMMC_EN pin to switch to an input with no pullup after a warm
reset, preventing the boot ROM from accessing the eMMC.

Change the pad mux for the eMMC pin to always keep the pullup
enabled.

Change-Id: I673a07e8f58fbea3c809fe5394655b3d43548cc6
Signed-off-by: Colin Cross <ccross@android.com>