input: mms_ts: ensure that we only enable/disable once
If the input device is closed after the early_suspend handler
turns off the controller and irq, the irq would have been disabled
a second time. Add a flag to keep track of whether the device is
on/off. When we remove the early_suspend handlers, this should
no longer be necessary.

Change-Id: I19ebd4a70a4d04e743a2b5894cb311e9f3223c21
Signed-off-by: Dima Zavin <dima@android.com>