power: max17040: Fix crash on race scheduling worker at probe
Initialize the worker before calling power_supply_register().  The
external_power_changed callback may be called immediately by
power_supply, which will schedule the worker, prior to initializing
the worker.  The uninitialized data structure can trigger a BUG in
the workqueue code, causing the worker thread to die.

Also move the reset prior to power_supply_register(), to ensure
the worker isn't run prior to, or concurrent with, the reset.

Change-Id: I969cc04dc6eb4579b8b91f788383addeb7f8d129
Signed-off-by: Todd Poynor <toddpoynor@google.com>