OMAP4: Tuna: Fix DSI Interleaving CIO Bus Contention Errors
There is a BW crunch while sending the second BTA from the panel
PHY in LP-RX mode) to OMAP, to put back the OMAP-DSI module in HS mode.
Because of a lesser low power clock, the request was NOT completed
on time before VIDEO_MODE start sending short sync packets.
Hence contention errors.

Increased the LP clock as the fix.

Change-Id: I37f86267170aad72c4533e9b709008e7d1afcf7e
Signed-off-by: Mark Tyler <mark.tyler@ti.com>
Signed-off-by: Lajos Molnar <molnar@ti.com>