misc: fsa9480: USB OTG support
The FSA9480 has a bug that prevents it from detecting a
cable detach when a USB OTG cable is connected.  To
workaround this the board file passes in a GPIO that the
FAS9480 driver can use to check the state of the ID signal.
The FSA9480 driver sets up an interrupt to catch the ID
signal going high when the cable is detached.  On a USB
host cable detach the FSA9480 will be reset to allow it to
detect the next cable change.

Change-Id: I0e7dd07682f056b3455dc9ab03e89d5fdeb16587
Signed-off-by: Adam Hampson <ahampson@sta.samsung.com>