ASoC: Disable pulls on WM8994 AIF2 when starting it
Pull control is availalbe for WM8994 AIF2, generally disabled as part of
the GPIO configuration in order to save power after system startup. As on
newer devices in the series there is no GPIO functionality on these pins
this will happen less naturally so have the driver disable the pulls as the
AIF is probed.

Change-Id: I5c77c27aac0f3ad4edde3ff544cd126c7f1959b3
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>