OMAP: DSS: panel-s6e8aa0: fix locking around panel start/stop
Moved the s6->lock mutex taking to outside of start/stop since the
dssdev->state variable was being modified without the lock held. The
backlight driver is relying on this state to know when the panel
is active so that it is safe to send commands over the MIPI link.

Change-Id: Idf2ea9547203829e7b18a272601f3f52a5d31cf5
Signed-off-by: Dima Zavin <dima@android.com>