power: max17040: add support for scaling the reported capacity
The battery capacity reported may not always be representative of
how much charge is actually remaining. When the battery is almost empty,
the output voltage may dip and go below the threshold for safe operation.
The "almost empty" watermark will be different for different designs so it
needs to be passed through platform data.

The battery reported capacity is thus scaled from [min,~100] to [0,100].

Change-Id: I257c67bcc9b47d833f0bcea9831f6c4a4486a58b
Signed-off-by: Dima Zavin <dima@android.com>