ARM: omap4: tuna: increase vmalloc space to 768MB
The default vmalloc space of 128MB results in 768MB of memory
used as lowmem, which is a waste of address space.  Most memory
allocations go to userspace, which doesn't care if the memory
came from lowmem or highmem, and allocates pages individually,
which will fragment lowmem.  Increase the vmalloc space to 768MB,
which will reduce lowmem to 128 MB.

Change-Id: Icd5ab94d5e93f627f93000bf6b44c78e3a1071e1
Signed-off-by: Colin Cross <ccross@android.com>