Revert "ARM: omap: tuna: Set OMAP_WAKEUP_EN to wlan irq pin"
This reverts commit 96eedbdd43bdb1bbe5cefb6c02d1fdb7b3ed844c.

The wlan wakeup issue is fixed by setting the GPIO as edge
detect instead of level only. That is addressed on the
previous GPIO patches. We don't need to set pad wakeup for
GPIOs that are in wakeup power domain. These GPIOs should
wake us up and not the control PAD.

Change-Id: I199914fb11434c7144b886165676582b48678114

Signed-off-by: Axel Haslam <axelhaslam@ti.com>