OMAP: DSS: panel-s6e8aa0: Move acl_set attribute to bldev and add acl_average
Move the acl_set attibute so it is in the same directory as the
brightness attribute.
Add an acl_average attribute to change the Y-average-accumulation-count
register value and a pdata argument to set the default value.
Also remove attributes if the device goes away and remove unused
bl_acl_off pdata value.

Change-Id: Ie74e35a307cda61a71b45fc13216ad63f578256f
Signed-off-by: Arve Hjønnevåg <arve@android.com>