ARM: omap4: tuna: Fix kernel panic when booting with JIG on factory line
To verify the version info of production binary after downloading,
we must use a JIGBOX switched to uart. But at that time factory program controls
JIG setting uart attach/detach action is happend during kernel booting.
But tuna_ap_usb_detach() is executed because there's no case for the UART.
So when detaching the UART we just skip the usb detach action.

Change-Id: I0de4ecd062dd298dbaaf23699702fa5f43524027
Signed-off-by: Sang-Jun Park <sj2202.park@samsung.com>