ARM: omap: tuna: enable/disable clk32kaudio for wlan
Wlan needs the clk32kaudio clk to be enabled for it
to function in low power modes. It has been observed
that during OFF mode if this clk is turned off, WLAN
will not wake up. We use the regulator framework to
enable and disable the clk32kaudio clk when WLAN is
turned ON/OFF.

Change-Id: I6662a67055000b167ec32bdd4df2ed9304a0a687
Signed-off-by: Axel Haslam <axelhaslam@ti.com>