ARM: omap4: tuna: hold clk32kaudio for BT
BlueTooth needs the clk32kaudio clk to be on
to be able to work properly. This clock may go
off in OFF in off mode, and when the clk32kaudio clk
was introduce as a resource may go off when wlan
is switched off.

Turn on clk32kaudio clk if BT is on.

Change-Id: I1f4a5202978112eb49da0aa8035047cf73db5fc8
Signed-off-by: Axel Haslam <axelhaslam@ti.com>