modem_if: usb_link: fix race of suspend and rx-submit
An urb completion can happen in middle of if_usb_suspend.
For such an URB, status == 0 and should not be re-submitted since
usb_kill_anchored_urbs() is in the process of killing all 'reading' URBs.

So check if suspend handler has alredy been called, then put the rx urb in
the 'urbs' anchor so that it can be resumed on if_usb_resume.

This race leads to error of wrong address going to EHCI and resulting in:
[  232.273712] WARNING: at arch/arm/mach-omap2/omap_l3_noc.c:129 l3_interrupt_handler+0xe4/0x380()
[  232.284118] L3 standard error: SOURCE:ABE at address 0xd20

Change-Id: I6013fd0ecd532805a55ef4f5c24181c67dad9451
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>