ARM: omap4: tuna: reclaim 32MB of memory
Adjust carveout sizes to the amounts of memory we actually need.  Reduce
the nonsecure tiler carveout by 17MB and the 2D carveout by 15MB.  The
nonsecure tiler carveout handles 2D surfaces for the GPU and is separate
from the 2D carveout, because the latter is firewalled during secure playback,
and the GPU cannot access the memory in this case, but still needs to
render nonsecure elements. Since the nonsecure tiler carveout is now 15MB,
we reduce the secure 2D carveout by this amount.

Change-Id: Iaf838348c9cb8c4ed193a4e568dc1d5fe1168c63
Signed-off-by: Iliyan Malchev <malchev@google.com>