ASoC: Implement support for WM1811A jack detection
The WM1811A features an advanced low power accessory detection subsystem
which allows the device to be maintained in a very low power state while
the system is idle without sacrificing any accessory detection features.

Implement software support for this, automatically managing the power
configuration of the device depending on the detected accessory.

Change-Id: I33692aa00cde6fc9a31e97786a2e94404c7019c6
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>