btwilink: avoid double freeing of recv skb
Upon error in hci_recv_frame, the HCI Core is freed before the error
is reported to btwilink. When the error does get to btwilink and forwarded
to ST core, ST core also attempts to free the skb.

Although the 2 frees of skb should theoritically cause any harm, this is
redundant and hence this patch returns 0 to ST core even upon error,
because there is no action to be taken.

Remove also the #define DEBUG which was used to debug the code.

Change-Id: I05f8c3a9d9795ac3e3e9916006630e809d27e33a
Signed-off-by: Manjunatha Halli <x0130808@ti.com>
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>