OMAP4+: HSI: Aligned HSR FSM control onto CAWAKE detection
HSR FSM enables the ACREADY to go to 1 and allows reception.
With previous code, HSR FSM is enabled/disabled when clocks are enabled/disabled.
But HSR FSM control shall be separated from HST state, and depend only on CAWAKE.

Now HSR FSM is disabled/enabled when CAWAKE is low/high.

Change-Id: I527bc0cc376218abd01d4e1c6cda2487ce0ecf27
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>