ASoC: wm8994: Support external capacitors on MICBIAS2 with jack detection
When an external capacitor is connected to MICBIAS2 on devices with
jack detection (which is not required but may be done in some systems)
then the loading may mean that better performance is obtained when
the microphone bias is enabled normally rather than using the low power
mode. Provide platform data allowing systems to indicate if they require
this.

Change-Id: I1a5a23e7ce3df09f7250910cef73eb773001e9a2
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>