OMAPDSS: DISPC: Allow -1 row-pixel increment on OMAP4/5
DSS used to throw a BUG on pixinc for pixels == 0.  However, this
is supported on OMAP4/5 where predecimation is enabled, e.g. if
we are displaying the full input image and

screen_width % x_decim == (x_decim - 1)

Also simplified if-else conditions that in effect calculated the
same equation.

Change-Id: I44047b677a84a05bdcfb53cfcb8b42f535a560df
Signed-off-by: Lajos Molnar <lajos@ti.com>
Signed-off-by: Sreenidhi Koti <sreenidhi@ti.com>
Signed-off-by: Varadarajan, Charulatha <charu@ti.com>
FIXUP from B&N OSS 2.2.0: proper check for UV feature
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>