lib: add generic strnlen_user() function
This adds a new generic optimized strnlen_user() function that uses the
<asm/word-at-a-time.h> infrastructure to portably do efficient string
handling.

In many ways, strnlen is much simpler than strncpy, and in particular we
can always pre-align the words we load from memory.  That means that all
the worries about alignment etc are a non-issue, so this one can easily
be used on any architecture.  You obviously do have to do the
appropriate word-at-a-time.h macros.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Keep module.h, since the export.h split isn't in K3.0
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>

Change-Id: I410a3ee8b98bdbff4a36f990d9f86b699cdc88e6