f2fs: add a sysfs entry to control max_discards
If frequent small discards are issued to the device, the performance would
be degraded significantly.
So, this patch adds a sysfs entry to control the number of discards to be
issued during a checkpoint procedure.

By default, f2fs does not issue any small discards, which means max_discards
is zero.

Change-Id: I080d38950d2bb35d9d0fc141e9d28921bf932c84
Signed-off-by: Jaegeuk Kim <jaegeuk.kim@samsung.com>