Ducati reloading fix for early suspend.
When the device is suspended with the running camera and then resumed,
it's possible that autosuspend counter won't reach zero and the main
condition in rproc_last_busy() won't be satisfied, so rproc won't be
restarted.

Fixed that by tracking the need to restart rproc separately and
independent of pm state.

Also cleaned up reference counting in rproc_put and rproc_get a bit
to make sure that forced offline and restart don't screw up the number
of references + won't result in race conditions.

Change-Id: I0f049632ca141d7d7447ce63a6022b81fcb1fee6