espresso: wifi: calculate MAC address based on the die ID
The Samsung BCMDHD used to read the factory MAC address from /efs. Accessing the
filesystem in-kernel is a very bad practice, but what can we expect from Samsung?

This commit calculates the MAC address based on the SoC Die ID: this is unique
for all devices, and ensures that the same device will always have the same MAC
address, not a randomly generated one.

Change-Id: Ie62935f79149e82c62462828880de2610b93ee5f